Get a basic parser working
So the basic idea is that the Parser receives the contents from a file.
Once it has its contents it can then iterate over its contents reading
up until the next token value.

Really the Parser just pushes and pops stuff off of the stack and it
manipulates these operations to find the "values" of the YAML file.

Ideally... each "Value" should have a "Key" associated with it. For
instance, if we have a "Key" we could assign it to the Token like so:

```
enum Token {
    // ...

    StringValue(String, String),
    ArrayValue(String, Vec<String>),

    // ...
}
```

This way the parser is able to hand back something that can be
translated into a `HashMap<String, Value>`.

Right now it treats keys as their own `Token` which is making things a
little bit annoying.